+++ Example/src/page/TabBar/TabBarItem/index.js
@@ -12,7 +12,7 @@ import PropTypes from 'prop-types';
 
 const styles = StyleSheet.create({
   all: {
-    height: 40,
+    flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
   },