+++ Progress/index.js
@@ -1,5 +1,6 @@
 import React, {
   Component,
+  PropTypes,
 } from 'react';
 import {
   StyleSheet,
@@ -7,7 +8,6 @@ import {
   Animated,
   Text,
 } from 'react-native';
-import PropTypes from 'prop-types';
 import Easing from 'Easing';
 
 const styles = StyleSheet.create({