+++ CHANGELOG.md
@@ -14,6 +14,8 @@ n3：💚 不影响之前版本使用方式的更新（不需要用户适配）
 
 ### TabBar
 
+- 💛 TabBar items will divide space equally and the touchable area will be extended as far as possible. You may need to add `flex: 1` in style of TabBar item components to adapte to this change.
+
 ## 0.18.0
 
 - 💚 import `PropTypes` from `prop-types` instead of `react`