+++ CHANGELOG.md
@@ -12,6 +12,14 @@ n3：💚 不影响之前版本使用方式的更新（不需要用户适配）
 
 - 💚 new props `hitSlop`：defines how far a touch event can start away from buttons in left and right
 
+### SmsCaptchaInput
+
+- 💚 new props `btnTextNumberOfLines`：used to truncate the button's text with an ellipsis after computing the text layout
+
+### TabBar
+
+- 💛 TabBar items will divide space equally and the touchable area will be extended as far as possible. You may need to add `flex: 1` in style of TabBar item components to adapte to this change.
+
 ## 0.18.0
 
 - 💚 import `PropTypes` from `prop-types` instead of `react`