+++ CHANGELOG.md
@@ -7,10 +7,11 @@ n1：❤️ 重大更新
 n2：💛 影响之前版本使用方式的更新（需要用户适配）
 n3：💚 不影响之前版本使用方式的更新（不需要用户适配）
 ```
+## v0.17.4-patch-2
 
 - 💚 new component for displaying progress：`Progress`
 
-## v0.17.5-patch
+## v0.17.4-patch-1
 
 ### NavBar